Make repo_info independent of sharding setup
Summary:
Apparently, in `SCMQuery` we rely on the `SCS` `repo_info` method to determine if the repo is being served by `SCS` and the commit_identity used by the repo. During conveyor pushes, we sometimes get situations where shards are temporarily displaced, i.e. they do not get allocated to the servers during restart. In such a situation, we might get `REPO_NOT_FOUND` error when hitting the SCS `repo_info` endpoint. This can make SCMQuery consider the repo to be backed by SCS by default.

Liu already made changes on SCMQuery end to avoid this incorrect assumption, but we should also fix the problem from SCS end. A method like `repo_info` should be servable for any repo regardless of which repos are actually present on the server. This diff does exactly that so the response for this method is not dependent on conveyor rollouts or rolling restarts.

Note that this does not mean this method will NEVER see `REPO_NOT_FOUND` error. If you request a repo that is gibberish or a repo that is valid but not in the SCS config tier, you can still get this error message.

Differential Revision: D76727698

fbshipit-source-id: 06354e474ed90882d9f274c4a8b0b18c62d56cc9