Introduce typed error for large repo not found
Summary: I intend to modify the `repo_info` SCS method to be independent of the sharding status of the server, i.e. even if the repo is deep sharded or shallow sharded `repo_info` returns the right information about the repo. For that, I need a typed error for the `Large repo not found` error message.  The reason for this will become apparent in the next diff

Differential Revision: D76726982

fbshipit-source-id: 092bcc0bf0ae2eb75b195132ef04cfc74f1fc906