gitcompat: do not read {FETCH,MERGE}_HEAD as loose ref files
Summary:
The `FETCH_HEAD` is different from `HEAD` in a way that it can be empty, or
contain multiple entries, like:

  515fd3a3b8d7a3cde2e542edaa5eefe952aed224	not-for-merge	branch 'jch' of https://github.com/git/git
  f93ff170b93a1782659637824b25923245ac9dd1	not-for-merge	branch 'maint' of https://github.com/git/git
  e2067b49ecaef9b7f51a17ce251f9207f72ef52d	not-for-merge	branch 'master' of https://github.com/git/git
  4767266eb4cedeaf40e4bda87fb9c8b6ae1ef127	not-for-merge	branch 'next' of https://github.com/git/git
  99452fd9cab409ea8cac682793f7065e3015a198	not-for-merge	branch 'seen' of https://github.com/git/git
  9a307cd2d656df87bcab0772285a3602af3e9a42	not-for-merge	branch 'todo' of https://github.com/git/git
  6768cd3982a5cf8df192224d4da8c3f03d4cb964	not-for-merge	tag 'v2.48.1' of https://github.com/git/git
  6ac5b341b89be0dbabb8780a8e89d1bb59dcb535	not-for-merge	tag 'v2.40.4' of https://github.com/git/git
  7e122199d2fac08cec62e6f9ef0edec53075c426	not-for-merge	tag 'v2.41.3' of https://github.com/git/git

It is not (yet?) used in our codebase either, so let's just do not read it as a
loose reference.

Related: https://github.com/facebook/sapling/issues/1007. The report suggests
an empty `FETCH_HEAD` could be an issue. However, `FETCH_HEAD` is filtered by
the `GitRefPreliminaryMatcher` and I didn't reproduce the empty `FETCH_HEAD`
breakage. I suspect the actual issue users saw is `*.lock` ref files, which
should be fixed by D76604157. Anyway, it seems harmless to stop reading
`FETCH_HEAD`.

Reviewed By: zzl0

Differential Revision: D76632052

fbshipit-source-id: 9620907751716998f96bd90b9b56bf7018eae05c