read from master when fetching the latest bundle list
Summary:
We read when we've generated bundle for a repo for the last time, and based on that we calculate how long we should sleep before attempting it again. This doesn't work as expected when we have no read-after-write consistency - we end up reading an old bundle so we regenerate bundle for a repo unnecessarily 2 times in a row.

This is the code doing it:
https://www.internalfb.com/code/fbsource/[212ea13667e529faf038e28b1815c296ccf647f0]/fbcode/eden/mononoke/facebook/git_bundle_generator/src/generator.rs?lines=344-377

We're reading from master here to get around this problem and get the read-after-write consistency.
The DB is very low traffic so this should be fine.

Reviewed By: RajivTS

Differential Revision: D76738564

fbshipit-source-id: c8ddea7585aabf3c18b02b8a541fd47147c54ac2