commands: wait for signal handler before propagating command error
Summary:
When interrupted, Sapling runs some "at exit" cleanup handlers in a different thread and then exits (in that thread). During that time, the normal command codepath is still running. If the cleanup work crosses wires with the command (e.g. trying to remove the incomplete clone directory during an interrupted "clone" command), the command codepath can end up propagating a "random" error that exits the command before the signal handler thread finishes cleaning up.

Add a sleep so that the main command thread "yields" to the signal handling thread. If the signal handler hasn't exited after a few seconds, the main thread proceeds (this should be very rare).

Reviewed By: MichaelCuevas

Differential Revision: D76630584

fbshipit-source-id: 0361a9c5ef9fc90c3259fe6671c414aff6c0955a