clone: clean up repo when interrupted
Summary:
Set up an atexit handler so we delete the incomplete repo when killed via a signal.

We don't support cancelation, so while we are trying to clean up the repo, the main thread is still trying to perform the clone. I threw in a bunch of retries to make relatively sure we clean up the repo.

Reviewed By: quark-zju

Differential Revision: D76630585

fbshipit-source-id: be943df0def89e134c08ff82f6fd291f123c08f0