walkdetector: retain metadata about large dirs indefinitely
Summary:
Keep directory entries with known large sizes around in the tree structure indefinitely. This way we know they are large directories if there is walk activity later.

Knowing that a directory is large is important to avoid false positive walk detections. If we incorrectly say a big directory is being walked, we will start prefetching potentially a lot of data unnecessarily.

This will use a small bit of memory, but directories with hundreds of entries are relatively uncommon, so it shouldn't be noticeable.

Reviewed By: zzl0

Differential Revision: D76647628

fbshipit-source-id: c8e55a8975e5ab8dbf7dec350702c3e557c44dfe