mononoke_api: rename fields of ChangesetPathDiffContext
Summary:
`ChangesetPathDiffContext` represents the difference between two paths in the repo (or the same path at two versions), and "old" version and a "new" version.

The current names, "base" and "other" are confusing.  Typically, "other" is the parent of "base", which conflicts with normal commit nomenclature where the first commit in a stack is the "base".

Change instead to explictly call these `new_content` and `old_content`.  This matches what the metadata diff function was already calling them.

Reviewed By: lmvasquezg

Differential Revision: D76620072

fbshipit-source-id: a8ea2e14a1f5f902f9964f664de818204e9238c3