cas: replace thin client with bare rust thrift client
Summary:
Context: we're replacing the thin cas client with a bare rust thrift client. This is to get rid of the start up overhead from creating a service router client and to improve reliability by getting rid of the C++ dependencies that the external CAS client pull in.

Replace the thin client with the rust bare thrift client.

Reviewed By: liubov-dmitrieva

Differential Revision: D76593861

fbshipit-source-id: 13d783ab1bee52baed02d0d580a12283f62744f6