fix a bug in handeling a variable in lambda function
Summary:
## Context
There is an issue with the way that `inodeNumber` was handling in `Nfsd3ServerProcessor::dispatchRpc` function. Actually `inodeNumber` is passing by reference here. This can lead to problems because `inodeNumber` is a local variable that will go out of scope when the `dispatchRpc` function returns. If the lambda function is executed after `dispatchRpc` has returned, `inodeNumber` will be invalid, leading to undefined behavior and eden crash.

This is the root cause of the crash which was introduced in `20250612-180204`.  [Scuba link](https://fburl.com/scuba/macos_crash_tailer/0je9zcan)
I contacted with a few users who experienced this crash, and collect some logs: P1844101840, P1844101915, P1844149923, and P1844149295

All of the crash happened when there is an issue and the requests are slow. Usually before all of the crashes we can see a few `Timeout was reached` in those logs. It match with the fact that when the request is slow the `.thenValue` lambda function run later, and with the time that eden gets to the `.thenValue` the main `dispatchRpc` function returns and its local variables are out of scope then we got the eden crashes.

## This diff
Capture `inodeNumber` by value instead of by reference to fix this issue. In this way, the `inodeNumber` lifetime extend to the end of `.thenValue` lambda function.

This was the only change in `20250612-180204`, and this crash begins when this release gets to prod. Therefore, I am pretty sure this is related to the crash and can fix this crash.

Reviewed By: MichaelCuevas

Differential Revision: D76856212

fbshipit-source-id: c09119d3314f3f4fec3d863758779379ee03ac97