Implement fix for pushing branches referencing tag in Mononoke Git server
Summary:
At some point during the push path development, I realized that users can push namespaced tags. These are tags where the name recorded in the tag object does not match the actual ref name pushed by the Git client. e.g. If a user pushes a namespaced tag with name `refs/namespaces/foo/refs/tags/some_tag` the actual name recorded in the tag object is `refs/tags/some_tag`.

Due to this discrepancy, we decided to rely on the pushed ref name to identify the right name to be used in bonsai tag mapping. However, then we discovered the use-case in D76820648. This meant that the pushed ref might have nothing to do with the actual annotated tag but might just be using as a reference for creating a branch pointing to the same commit. Our logic of using the pushed ref name would result in such refs getting stored as tag in our DB. This diff fixes that problem

Differential Revision: D76835259

fbshipit-source-id: 59e66972b2e9edef3d5c43bc3a3581becc4b90ae