Log ids of all blobstores that have reads disabled
Summary:
Generalize the changes I made in D76728908 to log if reads were disabled for the XDB blobstore.

This will check the JK for blobstore ids 1 to 5 and log which ones are disabled (if any).

## Why this change?

We'll soon run another experiment where we'll shut down the second manifold blobstore for some requests. Instead of adding another column, I figured I would just add a generic one to store all the ids.

## RFC
Should I check ids 4 and 5 or is it not worth it?  I know JK checks are no expensive, but if in practice we're only using 1 to 3, it's probably worth changing the range. clara-9 markbt

Reviewed By: clara-9

Differential Revision: D76895703

fbshipit-source-id: c945c817647bb0ff3a57282b1a627142ac932cc3