Fetch git objects from the blobstore as `ObjectContent` instead of owned `Object`s
Summary:
For round-tripping, we need to replace all instances of owned gitoxide types to ObjectRef.
When ownership is actually needed, ObjectContent is the correct type to use as it stores the owned bytes in addition to the semantic ObjectRef.

Update the output of `fetch_non_blob_git_object` and `fetch_git_object` to return `ObjectContent` and propagate to the call sites.

Reviewed By: RajivTS

Differential Revision: D76809131

fbshipit-source-id: 11b6feb6ac2ed07d673c674c92263336d74d7536