mononoke: expand capabilities to specify commit cloud and others
Summary:
Today `capabilities` reports `['commit-graph-segments']`. It assumes the "commit-graph-segments" is the only optional feature and everything else is available by default, which is no longer true for git repos.

For this we're adding two capabilities:
- `sapling-common` which indicates that most features like blame, megarepo, etc. are available
- `commit-cloud` which specifies that commit cloud relatd endpoints are supported

This was we can tell in sapling when is feasible to sue edenapi for certain operation and when it's not. In this diff, `mononoke/server` will report full capabilities which `git_server` will only report commit cloud as available.

Also given this is backwards compatible it allows us to do a faster rollout once client side assertion is implemented

Reviewed By: markbt

Differential Revision: D76892878

fbshipit-source-id: b98bd8810cf7d6bb56696815dd192da4594e9285