walkdetector: add walker pid tracking
Summary:
Now the Detector also accepts a process ID value when observing file/directory accesses. It propagates this into tracked walks so we can have some idea which process is doing the walking.

We don't do full tracking of paths, and I didn't want to track/handle multiple pids, so we instead take a simple probabilistic approach where we swap a walk's existing pid with the new in-hand pid based on how many file/dir accesses are associated with each. For example, the pid will swap with low probability when we see 1 file access by a new pid under an existing walk with 1_000_000 previous accesses, but with 50% when merging two walks together each with 1_000_000 accesses.

Reviewed By: zzl0

Differential Revision: D76874928

fbshipit-source-id: 568bc2b89ea24059966a984905813da656f1a2f9