debugshell: show source code in traceback when executing a file
Summary:
By passing the file path to `compile` instead of calling `exec` directly, the
traceback will include the source code.

The exception handling is now inside debugshell, so it won't trigger the
top-level error handling. It's no longer considered as a crash, and will
not print "Sapling has crashed..." message.

Reviewed By: zzl0

Differential Revision: D76917847

fbshipit-source-id: 218687dea17a9be96cfe628c6fd3fc2a0a31e19f