Retry data derivation for retriable errors everywhere
Summary:
Before this change, we had one place where we did retries and other places where we didn't.
By doing the retries inside `derive_all_types`, which is the common utility for deriving data in `megarepo_api`, we get consistent retries everywhere.
Remove the outer retry loop to avoid doing up-to 25 retries.

Also, use `retry::retry` to conveniently get exponential backoff and jitter.

Also retry only if the error contains "blobstore" or "timeout", which are retriable errors.
There is no point in retrying 5 times if we are simply passing bad data and the derivation should fail.
This can be adjusted if we still see derivation errors that are retriable in the future.

Differential Revision: D76978122

fbshipit-source-id: 54fc5019e07f11f1236d5ca6c53a107e30ad4e7d