fix how we set up repository root for get_status
Summary:
As of D76038889, `sapling_client::status::get_status` takes an `Option` parameter `repository_root` specifying the repository for which we want to get commit information.

Currently, this is achieved by passing `--repository` to `hg status` if `repository_root` is `Some`. The returned paths are relative to `repository_root`.
However, if the parameter `root` is also `Some` then we pass a `path:` filter to `hg status`, which causes the returned paths to be *relative to the current working directory*.

That's surprising to users of the function and should be made more consistent.

This diff changes the behavior so that if `repository_root` is `Some`, then the returned paths are always relative to `repository_root`.
This is achieved by changing `get_status` such that if `repository_root` is `Some`, we don't pass `--repository` to `hg status`, but instead we set the working directory for the spawned `hg` process to `repository_root`.

Differential Revision: D76974859

fbshipit-source-id: ee79f8eb4c6568b1fdd7becc065f348efe6bc072