Fix ISL diff views when using an external diff program
Summary:
The change in D75930834 broken some uses of external diff programs.  The extension has to create a temporary file for the external tool to read, and in order to not leak those files, it must delete them after the tool is done with them.  It allows the tool 5 seconds to open and read the file before deleting it.
Previously this was done with a Sleep command, but in D75930834 it was changed to `_ = Task.Delay(5000);`.  That call doesn't actually wait, however, it just creates an asynchronous task that waits.  We must await the task in order to actually wait.

Reviewed By: ViniGupta08

Differential Revision: D76929739

fbshipit-source-id: 16ee9b7680799b0f859d3fdb3a17e78ad7de3477