change the autosync threshold for the aux data
Summary:
change the autosync threshold for the file aux data

aux data is very small, however the autosync threshold was chosen `10 MB`
in-memory buffer which is large for this sort of data (for example, we use `50
MB` for file blobs and for LFS)

Imagine a scenario where Sapling prefetch is running in the background. This could lead to a delay before Eden detects the aux data in the cache. As a result, Eden may receive synced trees and files more quickly, but the aux data might not be immediately available, forcing it to either recalculate from cached content or even fetch from the remote source again.

I suggest setting the threshold to `1MB`, aligning it with the tree aux data sync threshold Mark has chosen. This consistency will simplify how we manage aux caches in Sapling due to their similarities, ultimately facilitating faster synchronization between Sapling and Eden.

```
lib/revisionstore/src/indexedlogtreeauxstore.rs
94:            .auto_sync_threshold(1_000_000)
```

Reviewed By: zzl0

Differential Revision: D77110590

fbshipit-source-id: 776e42f0e5e5cd4777cd91ac3507846fa5cc4282