refactor ConfigManager so that config fetching is implemented with a separate struct for each config source
Summary:
This diff refactors the different fetch methods into separate structs that implement a new "Config Fetcher" interface.

In this diff, the "fetcher" instance is still created and called within the ConfigManager.  The next step would to do that within main() and just pass the EdenFSConfig to the ConfigManager.  But first things first.

The trait isn't strictly necessary at this time, but I think it will make it easier for users of the library if they want/need a customized fetch policy for their particual use case.  For example, for Sandcastle, I'm thinking about using  a "virtual" fetcher that will call a sequence of fetchers to implement fallback behavior (e.g manifold then configerator).

Reviewed By: genevievehelsel

Differential Revision: D76771004

fbshipit-source-id: 38f569adfc64cdeff78a4d8e13c5760fca046275