refactor ConfigManager so that run() method takes a EdenFSConfig
Summary:
This diff refactors the ConfigManager so it's run() method takes an EdenFSConfig parameter instead of fetching configs itself.

This is to support users of the edenfs config manaager library to implement fetching strategies appropriate for their use cases.  For example, Sandcastle will attempt to fetch from Manifold (if a manifold key is specified by the job) and then fall back to Configerator.  None of the other fetch methods (once they are implemented) will be used.

Note that this ressulted in the `local_cfgr_root`, `canary_host`, and `manifold_key` fields being removed from the ConfigManagerOptions struct, as those values are used by the various "fetchers" and are no longer referenced from within ConfigManager.

In the future, I expect that fields like `raw_out` and `timeout` to be removed from ConfigManagerOptions as well (once the remote/interngraph client "fetcher" is implemented). Even now these fields aren't referenced by ConfigManager.

Reviewed By: genevievehelsel

Differential Revision: D76843944

fbshipit-source-id: 3edc1ef92e6513313f57940b7b7b1021911373b0