metaconfig: don't fail when encountering unknown fields in TOML config
Summary:
In prod we don't fail when parsing config that contains unknown fields. TOML config parsing in integration tests should mimic prod behaviour.

This wasn't an issue in the past but now we have trinity tests that test against a Mononoke package built from an older commit. In such cases having unknown fields in the config is expected, and testing that Mononoke works correctly after ignoring the fields is the correct prod matching behaviour.

Reviewed By: chiara125, liubov-dmitrieva

Differential Revision: D77175426

fbshipit-source-id: 82d7bba099bce485afd0a5a67e50e33c2876728f