Use tree and blob concurrency for shallow clones
Summary:
As part of shallow clones, we have to count the number of distinct objects in the shallow boundary requested by the client. If the shallow boundary contains lots of commits we might end up loading lots of trees for these commits and list the entries in those trees. The problem is exacerbated if the root trees are very wide. Both cases are true for repos like `aosp/platform/vendor/qcom/boot_images` which eventually result in OOMs with shallow clones.

I have already lowered the overall concurrency for these repos but the problem is we utilize the commit concurrency and not the trees and blobs concurrency for shallow clones. This diff fixes that.

Reviewed By: lmvasquezg

Differential Revision: D77168906

fbshipit-source-id: 250692fddede288b60213511996c03fd8a3c8838