Fix spelling in smaller directories in eden/
Summary:
eden/ is the biggest directory that uses the SPELL lint today:

https://www.internalfb.com/code/fbsource/[c85164c72fda58e31697071dafbbbb1ae5105b2f][blame]/tools/arcanist/lint/fbsource-lint-engine.toml?lines=4456-4461

I'm looking to move SPELL to using the `typos` tool, which catches a bunch more issues. This is the first diff fixing spelling issues in eden, focusing on the smaller directories (i.e. excluding monoke, fs, and scm).

Reviewed By: lmvasquezg

Differential Revision: D76983692

fbshipit-source-id: 770d93afd484556a2a926dbee2af60ce79df1745