derivation_queue: add unsafe_evict method
Summary:
Sometimes it's desirable to evict an item from the derivation queue without doing the actual derivation if there's something problematic about it (e.g. large or malformed commit). This diff adds an force evict method to the derivation queue to do so.

Since this method doesn't do the actual derivation it can cause dependent items to become stuck. This is reasonable since this method should only be used for items that can't whose derivation is stuck anyways.

Differential Revision: D77230413

fbshipit-source-id: 388669795f6740513c90300f8225055f99745d5b