admin/derivation_queue: add unsafe_evict subcommand
Summary: Add subcommand for unsafely evicting an item from the derivation queue. The command is unsafe because it can leave dependent items stuck in the queue since the actual derivation will not be performed. This is to be used to evict items that can't be derived normally for some reason (e.g. large or malformed commits).

Reviewed By: RajivTS

Differential Revision: D77231848

fbshipit-source-id: 2da2b5e3ef356515d38a7d2398e63380810e5eb7