PrivHelper: Add ProcessPriority stub
Summary:
# This diff

Memory priorities are just one aspect of process priorities that we may want to control. For example, CPU priority (i.e. nice value) and IO priority (i.e. ionice) could be controlled in the future. To leave an easy route for that in the future, we introduce a ProcessPriority class which allows us to store multiple priority types in a single platform-agnostic class.

For now, only MemoryPriority is included in the class definition.

# Context

S530151 is caused by frequent OOM kills on macOS. We also have frequent OOM kills on Linux. In many cases, EdenFS is not necessarily the cause of OOM kills. Instead, EdenFS is just caught in the crossfire and taken out because it uses a non-trivial amount of memory.

We can reduce the likelihood of EdenFS getting killed by raising it's memory priority. This stack aims to add logic into EdenFS to raise memory priority at startup time for both the daemon and the PrivHelper.

Differential Revision: D76936389

fbshipit-source-id: 7072fdbd4506cbba6bb2b4739afef49426295617