MemoryPriority: implement memory priority setting for Linux
Summary:
# This diff

Sets the oom_score_adj value for a given process by writing to the file under `/proc/<pid>/`. This is a privileged operation, but that's fine since the PrivHelper has permission to do so.

As far as I can tell, writing to this file directly is the best way to set the oom_score_adj. There is no other API/command for doing so.

# Context

S530151 is caused by frequent OOM kills on macOS. We also have frequent OOM kills on Linux. In many cases, EdenFS is not necessarily the cause of OOM kills. Instead, EdenFS is just caught in the crossfire and taken out because it uses a non-trivial amount of memory.

We can reduce the likelihood of EdenFS getting killed by raising it's memory priority. This stack aims to add logic into EdenFS to raise memory priority at startup time for both the daemon and the PrivHelper.

Differential Revision: D76948522

fbshipit-source-id: 15706fb02f7ba5ae36f4dea4c8513409d8501e03