MemoryPriority: add private kernel structs for setting memory priority on macOS
Summary:
# This diff

Setting Jetsam values is not an easy task. The only way to do so is via undocumented syscalls. To use these, we must define the signatures in a header value and mark them as "extern" so that the linker knows the implementation is defined elsewhere.

We also must include some structs that are used to call the syscalls for setting/getting Jetsam priorities.

Note: this code takes heavy inspiration from this GitHub project https://github.com/asdfugil/overb0ard/tree/nickchan

# Context

S530151 is caused by frequent OOM kills on macOS. We also have frequent OOM kills on Linux. In many cases, EdenFS is not necessarily the cause of OOM kills. Instead, EdenFS is just caught in the crossfire and taken out because it uses a non-trivial amount of memory.

We can reduce the likelihood of EdenFS getting killed by raising it's memory priority. This stack aims to add logic into EdenFS to raise memory priority at startup time for both the daemon and the PrivHelper.

Differential Revision: D76936383

fbshipit-source-id: 23354337b1de333ca6a281a5c558f543f1c79c19