Config: make daemon/privhelper memory priority configurable
Summary:
# This diff

Adds an EdenConfig that can be used to control the memory priorities of the EdenFS daemon and the PrivHelper. These are used later in this stack to gate (and set) memory priority logic.

# Context

S530151 is caused by frequent OOM kills on macOS. We also have frequent OOM kills on Linux. In many cases, EdenFS is not necessarily the cause of OOM kills. Instead, EdenFS is just caught in the crossfire and taken out because it uses a non-trivial amount of memory.

We can reduce the likelihood of EdenFS getting killed by raising it's memory priority. This stack aims to add logic into EdenFS to raise memory priority at startup time for both the daemon and the PrivHelper.

Differential Revision: D76936390

fbshipit-source-id: 1b5c84cf5921e0a55d0e4778b14dd8360bf595aa