Startup: set daemon memory priority
Summary:
# This diff

Modifies the daemon to set its own memory priority via the privhelper. The priority value is read from an EdenConfig. My current plan is to slowroll a high priority. Once we are sure that priorities don't cause issues, we can change the config to have a sane default value (200+ on macOS, -500 on Linux)

# Context

S530151 is caused by frequent OOM kills on macOS. We also have frequent OOM kills on Linux. In many cases, EdenFS is not necessarily the cause of OOM kills. Instead, EdenFS is just caught in the crossfire and taken out because it uses a non-trivial amount of memory.

We can reduce the likelihood of EdenFS getting killed by raising it's memory priority. This stack aims to add logic into EdenFS to raise memory priority at startup time for both the daemon and the PrivHelper.

Differential Revision: D76946800

fbshipit-source-id: bd536d1faf81feab3fda9cc91b2da912c1c62f27