ProcessPriority: move code to non-fb directory
Summary:
# This diff

This code was originally introduced under a FB-only directory. However, that just makes building harder to deal with. There's no reason for this code to be FB-only, so this diff moves it to an unhidden directory.

# Context

S530151 is caused by frequent OOM kills on macOS. We also have frequent OOM kills on Linux. In many cases, EdenFS is not necessarily the cause of OOM kills. Instead, EdenFS is just caught in the crossfire and taken out because it uses a non-trivial amount of memory.

We can reduce the likelihood of EdenFS getting killed by raising it's memory priority. This stack aims to add logic into EdenFS to raise memory priority at startup time for both the daemon and the PrivHelper.

Differential Revision: D76936387

fbshipit-source-id: d99f7b539704090098a552f49f2ca6ecac889fba