LinuxMemoryPriority: implement getPriorityForProcess
Summary:
# This diff

Reads the oom_score_adj value for a given process by reading it from the file under `/proc/<pid>/`. This is a privileged operation, but that's fine since the PrivHelper has permission to do so.

As far as I can tell, reading from this file directly is the best way to query the oom_score_adj value. There is no other API/command for doing so.

# Context

S530151 is caused by frequent OOM kills on macOS. We also have frequent OOM kills on Linux. In many cases, EdenFS is not necessarily the cause of OOM kills. Instead, EdenFS is just caught in the crossfire and taken out because it uses a non-trivial amount of memory.

We can reduce the likelihood of EdenFS getting killed by raising it's memory priority. This stack aims to add logic into EdenFS to raise memory priority at startup time for both the daemon and the PrivHelper.

Differential Revision: D76955318

fbshipit-source-id: 990b5a85efbd1aeee875215c792f6ec4ee872133