PrivHelper: add setMemoryPriorityForProcess requests
Summary:
# This diff

Since setting MemoryPriorities is a privileged operation, the PrivHelper must be the process that executes these operations. To allow this, we must add a new PrivHelper request type for setting memory priorities for a given process.

Since it's not crucial to EdenFS operation that memory priorities are set, we currently ignore failures (i.e. the PrivHelper request will return success regardless of the outcome). This is done so we can avoid adding compatability code during the intial rollout.

We can turn failures into actual request errors at a later date, if desired. Otherwise we can keep these requests as best effort

# Context

S530151 is caused by frequent OOM kills on macOS. We also have frequent OOM kills on Linux. In many cases, EdenFS is not necessarily the cause of OOM kills. Instead, EdenFS is just caught in the crossfire and taken out because it uses a non-trivial amount of memory.

We can reduce the likelihood of EdenFS getting killed by raising it's memory priority. This stack aims to add logic into EdenFS to raise memory priority at startup time for both the daemon and the PrivHelper.

Differential Revision: D76936388

fbshipit-source-id: d03622ba8569b75a2720bc1405cbd21649e2e95a