Allow Goto past landed revision/hashes
Summary:
[this one is for rev hashes, preceding diff is for diff numbers, which have a different problem]

Goto past/landed hashes doesn't work in ISL's "download commits" dialog (try using it to goto `rFBSb12a6c4e99479cbf1e19797a3d8283492ce5d1db`, for example).
Even with the "Goto" checkbox checked.

In this case, the `sl pull --rev` command will fail for past/landed commits that are already available locally (and `jf get`/`Internal.diffDownloadOperation` would probably also fail, if it got that far)

We can check if the revset is available locally and allow it to proceed if so, even if the pull commands fail, to complete the Goto.

Reviewed By: quark-zju

Differential Revision: D76999358

fbshipit-source-id: 2e62b19447d497090ba40d184ce20786a470b19f