Disallow void element in collection
Summary:
It is possible to trigger a DoS against a Rust Thrift service if the service's Thrift interface contains collections of a void element type, which is not natively expressible in Thrift, but is expressible using a `rust.Type` annotation:

```lang=thrift
rust.Type{name = "workaround::RustVoid"}
typedef i16 RustVoid

struct Arg {
  1: list<RustVoid> voids;
}

service Vulnerable {
  void repro(1: Arg arg);
}
```

where `workaround` is a Rust module included via `rust_include_srcs` or a Rust dependency via `rust_deps`.

```lang=rust
mod workaround {
    pub type RustVoid = ();
}
```

This is not a real use case we need to support, but it is worth preventing the vulnerability by detecting and returning a deserialization error in this case.

Reviewed By: cjlongoria

Differential Revision: D77315658

fbshipit-source-id: dff5baa4ed5d60667cfa43413d537565ecedfd68