Split out EdenapiSender setup to a function
Summary:
## This stack

The [recent modern sync breakage](https://fb.workplace.com/groups/1708850869939124/permalink/1997997347691140/) was caused by the fact we were unable to connect to fwdproxy. Root cause aside, this was hard to detect because we completed the shards setup but then failed to stay up. Conveyor cannot detect this.

The ideal fix is to make sure that we check all our dependencies work before we complete the shard setup.

## This diff

Split out the initialization so we can reuse elsewhere with minimal code duplication.

Reviewed By: lmvasquezg

Differential Revision: D77433160

fbshipit-source-id: 435e425c53bfd0d81bd864016f9e65f987b87978