Move context creation to a helper
Summary:
## This stack

The [recent modern sync breakage](https://fb.workplace.com/groups/1708850869939124/permalink/1997997347691140/) was caused by the fact we were unable to connect to fwdproxy. Root cause aside, this was hard to detect because we completed the shards setup but then failed to stay up. Conveyor cannot detect this.

The ideal fix is to make sure that we check all our dependencies work before we complete the shard setup.

## This diff

SSIA

Reviewed By: lmvasquezg

Differential Revision: D77441433

fbshipit-source-id: a5cdd1258bd87486937bfe885b6150c380f5f5ef