CwdSelector: integrate SubmoduleSelector into CwdSelector
Summary:
Let a submodule dropdown show up with CwdSelector as a breadcrumb.

Note that although it's possible to connect multiple `<SubmoduleSelector />` to reflect nested gitmodules, we're not really able to do so until either
* ISL able to preserve the inter-repo context; or
* sl able to return repo roots recursively.

My next step is to achieve that.

I want to publish this first for suggestion :) Tried my best to make the styling consistent with the current CwdSelector, but I don't have much experience in UI so far.

Reviewed By: evangrayk

Differential Revision: D77266601

fbshipit-source-id: 54bc6cfd353f6aa6f4285765d4554e8197fe5fb8