tests: add test to show merge failures caused by subtree copy commits
Summary:
The subtree-copy/modified conflict was added to prevent rebasing subtree-copy
commit onto a commit that also changed the to-path. Otherwise, the subtree-copy
will include unexpected changes from the dest side.

But for subtree merge, we should allow merging files for subtree-copy/modified
case. This diff adds a test to show the failures.

Reviewed By: markbt

Differential Revision: D77326926

fbshipit-source-id: 6271d5789ac932dd07e87968ef38601b539b159d