tests: fix merge failure caused by subtree copy commits
Summary:
The subtree-copy/modified conflict was added to prevent rebasing subtree-copy
commit onto a commit that also changed the to-path. Otherwise, the subtree-copy
will include unexpected changes from the dest side.

But for subtree merge, we should allow merging files for subtree-copy/modified
case. This diff fixes the merge failure.

Reviewed By: markbt

Differential Revision: D77326928

fbshipit-source-id: cf0bf09716b0756e23f551b0740d3b77dd72d9a7