Move sharded mode to its own command
Summary:
## This stack

The [recent modern sync breakage](https://fb.workplace.com/groups/1708850869939124/permalink/1997997347691140/) was caused by the fact we were unable to connect to fwdproxy. Root cause aside, this was hard to detect because we completed the shards setup but then failed to stay up. Conveyor cannot detect this.

The ideal fix is to make sure that we check all our dependencies work before we complete the shard setup.

## This diff

In the current setup we are passing an explicit `--repo-name opsfiles` argument to all shards, even if they handle other repos; the code ignores it, but that's error prone.

Let's split out sharding mode to its own command and *not* pass a repo name at all.

This diff only creates the new command but doesn't clean up the old one. Once a build is out with this, we'll be able to change the TW spec. Only then it will be safe to clean up the old code.

Reviewed By: lmvasquezg

Differential Revision: D77447723

fbshipit-source-id: 9df0420e8ec8cfb1bcf64075b013b4646f569bdc