Probe EdenapiSender during setup
Summary:
## This stack

The [recent modern sync breakage](https://fb.workplace.com/groups/1708850869939124/permalink/1997997347691140/) was caused by the fact we were unable to connect to fwdproxy. Root cause aside, this was hard to detect because we completed the shards setup but then failed to stay up. Conveyor cannot detect this.

The ideal fix is to make sure that we check all our dependencies work before we complete the shard setup.

## This diff

This will ensure that we can connect to AWS before we complete the setup. This is true whether we have anything to sync or not, which is particularly helpful when we're caught up.

Reviewed By: lmvasquezg

Differential Revision: D77433358

fbshipit-source-id: 4cdb5a7d4243aae63f1b9eba678b19495a878340