Move sync arguments down to the commands that need them
Summary:
## This stack

The [recent modern sync breakage](https://fb.workplace.com/groups/1708850869939124/permalink/1997997347691140/) was caused by the fact we were unable to connect to fwdproxy. Root cause aside, this was hard to detect because we completed the shards setup but then failed to stay up. Conveyor cannot detect this.

The ideal fix is to make sure that we check all our dependencies work before we complete the shard setup.

## This diff

I wasted close to an hour on D77433358 because I forgot that when running in sharded mode, we cannot use the repo args we get on the command line. That sucks and I want to undo it, so let's start by making sure only commands that do need these args, declare them.

Reviewed By: lmvasquezg

Differential Revision: D77443696

fbshipit-source-id: 5371add6d2745f66e867af9f190af76169968609