Add span fields when logging in test mode
Summary:
## This stack

The [recent modern sync breakage](https://fb.workplace.com/groups/1708850869939124/permalink/1997997347691140/) was caused by the fact we were unable to connect to fwdproxy. Root cause aside, this was hard to detect because we completed the shards setup but then failed to stay up. Conveyor cannot detect this.

The ideal fix is to make sure that we check all our dependencies work before we complete the shard setup.

## This diff

I wrote this mostly to validate D77446510; it might be too verbose to do by default. Accepting input on that.

Reviewed By: gustavoavena

Differential Revision: D77531241

fbshipit-source-id: 858a593c752e510320148accfd1b8969d1da9cf0