Ensure deletion of expired snapshots is possible
Summary:
We have a check that validates that running the update command to mark a snapshot as expired also returns non-zero number of modified rows. However, if the snapshot was previously marked as expired but not deleted and later if we try to delete that snapshot, we get an error because we get back zero modified rows from the update query.

Let's remove this check because I added this in my first few months in the company and was being super defensive in case I cause SEVs :)

Reviewed By: YousefSalama

Differential Revision: D77594133

fbshipit-source-id: 23812d51e2d36611d7db9f083ffbd6ca59fe8ad5