Fix spelling in eden/scm/ (eden 3/4)
Summary:
eden/ is the biggest directory that uses the SPELL lint today:

https://www.internalfb.com/code/fbsource/[c85164c72fda58e31697071dafbbbb1ae5105b2f][blame]/tools/arcanist/lint/fbsource-lint-engine.toml?lines=4456-4461

I'm looking to move SPELL to using the `typos` tool, which catches a bunch more issues. This is the third diff (of four) fixing spelling issues in eden, this time focusing on the `eden/scm/` subdirectory.

ignore-conflict-markers
fbcode/eden/scm/contrib/smerge_benchmark.py has intentional markers in a doc comment.

Reviewed By: quark-zju

Differential Revision: D77315200

fbshipit-source-id: 0e40da833ad58723b12719ee98d402b9345feb45