Be lenient when importing timestamps from Git
Summary:
Originally, we failed if the date in a Bonsai commit could not represent a Git date.

This is not necessary anymore as we store the raw bytes from Git during import and can roundtrip from Git to Git without serializing and de-serializing into a Bonsai.

It is also harmful as it is blocking the import of `aosp/abl/tianocore/edk2`.

Git dates are more lenient than Mononoke's, which means we can always derive a Git Commit from a Bonsai.

Clamp the timestamp offset to valid values during conversion from Git dates to Mononoke dates.

Reviewed By: YousefSalama

Differential Revision: D77596897

fbshipit-source-id: bd07f6f1101fa9aafd001828acf1b8eff22d0b3e