Enable controlled concurrency for snapshot deletion
Summary: When I joined the company I knew no Rust so ended up doing a `try_join_all` on the futures for deleting the snapshots. That was assuming that we wouldn't want to delete more than a reasonable number of snapshots at a time. However, that isn't the case anymore so let's use streams with user-controlled concurrency so we can control how many we delete at a time

Reviewed By: markbt

Differential Revision: D77654734

fbshipit-source-id: a6c7a627095a46b50dba70d65ede59af2e47f0be