tests: add a test to revert a subtree merge
Summary:
## Context

Users may want to revert a subtree merge commit if it contains a bad code change.
However, if they use sl backout, the next subtree merge may mistakenly treat the
backout as a conflict resolution, causing it to skip the source-side changes.

We are planning to use "subtree copy --force" to overwrite the dest side and
make merge-base follow the subtree-copy source to skip the subtree-merge
commit.

## This diff

This diff adds a test demonstrating that the current merge-base search does
not follow the subtree copy overwrite source.

Reviewed By: quark-zju

Differential Revision: D77326929

fbshipit-source-id: fd5f3ef3764fe099df4f9d410226830831daeff0