hgmain: trim leading ';' off windows $PATH
Summary:
Fix $PATH with leading ';' to work on Windows. This fixes sl to be able to search the $PATH to find, for example, the editor "notepad" (when configured without absolute path).

I've discovered that setting NoDefaultCurrentDirectoryInExePath (see D42154134) has a side effect of not cleaning leading ';' form $PATH. I'm assuming the cleaning happens as a side effect of prepending "." to $PATH. Fix by manually stripping leading ';'.

Reviewed By: zzl0

Differential Revision: D77672892

fbshipit-source-id: 3316aff4cd61b7c70d0061865af84f327f462135