Cleanup after moving sharded mode to its own command
Summary:
## This stack

The [recent modern sync breakage](https://fb.workplace.com/groups/1708850869939124/permalink/1997997347691140/) was caused by the fact we were unable to connect to fwdproxy. Root cause aside, this was hard to detect because we completed the shards setup but then failed to stay up. Conveyor cannot detect this.

The ideal fix is to make sure that we check all our dependencies work before we complete the shard setup.

## This diff

D77447723 introduced a new command to run in sharded mode. Once the TW job switches, we can remove support from `sync-loop` to avoid any confusion.

Reviewed By: lmvasquezg

Differential Revision: D77448901

fbshipit-source-id: 0b843e629af427a1307df71effd9f1ef6d6b55f0