Migrate logtools/categorizer targets
Summary:
# This diff
This diff updates a bunch of categorizers using the deprecated Thrift libraries. Some of them are easy to migrate since they use the Py3 version, while others are harder to migrate because they use the Python-deprecated version.

# This stack
This diff stack is part of addressing T225293165 and T225293675 which requires migrating `thrift-py-deprecated` to `thrift-python`.
See wiki pages for migration guidance:
- [Main Page](https://www.internalfb.com/wiki/Thrift-Python/)
- [Serialization](https://www.internalfb.com/wiki/Thrift-Python/User_Guide/Serialization/)
- [Quick Migration Guide](https://www.internalfb.com/wiki/Thrift-Python/Migration_Guide/Quick_Migration_Guide/)

Reviewed By: markhua

Differential Revision: D77703295

fbshipit-source-id: 5d931e88c8fc33a837ba565bc4e0f9ea2157cd39