Start logging some telemetry to Scuba
Summary:
## This stack
Previous diffs focused on piping the stats returned by the MySQL client up to mononoke_queries.

The next few diffs will focus on changing mononoke_queries and the mononoke codebase to start logging the data.

## This diff
With the data available, let's start logging some of it to Scuba.
For now, we'll not be logging any mononoke-specific data (e.g. repo, service), but I'll be adding that later.

This will only log the MySQL stats.
I intent to add
- Mononoke things: repo, repo id scheme, service, etc
- Full query, so we can debug it later.

## Notes
- The way I'm setting the Scuba table is not great. I did a quick hack to be able to override the destination to a file in integration tests, but I intend to improve that later.
  - I'll start my oncall later today and I want to start logging something before that.
  - After we examine the data, I'll iterate on it and plan to use a schematized logger, so this hack will be removed anyway.
- I'm not sure if I need to worry about sampling here, or if data will be sampled by default in a new scuba table... markbt any idea?

Reviewed By: markbt

Differential Revision: D77660561

fbshipit-source-id: d0b84566781793968db7eb4c4a3aed9a471c1574