Update all callsites that pass ClientRequestInfo to pass CoreContext
Summary:
## This stack
Previous diffs focused on piping the stats returned by the MySQL client up to mononoke_queries.

The next few diffs will focus on changing mononoke_queries and the mononoke codebase to start logging the data.

## This diff
Remove a temporary `From` implementation that I used with `ClientRequestInfo` and make sure all of the callsites that were using it pass a `CoreContext`.

Reviewed By: lmvasquezg

Differential Revision: D77726107

fbshipit-source-id: 88a9e8335018b81ac99568f449211378e4fc06d5